Opg3_Fix

+++ SPEC_Uge3/opg3/opg3.py
@@ -1,5 +1,5 @@
-import io
 import os
+import sys
 
 try:
     file = open("SPEC_Uge3/opg3/source_data (Datamigrering hjælpeværktøj).csv","r")
@@ -7,24 +7,26 @@ try:
     file.close()
 except (IOError, FileNotFoundError) as e:
     if type(e) == IOError:
-        print("Der er fejl")
+        sys.exit("Læsningen af filen fejlede")
     else: 
-        print("Der er en anden fejl")
+        sys.exit("Filen blev ikke fundet, tjek om filen er i denne mappe")
         
-    
 
 script_dir = os.path.dirname(os.path.abspath(__file__))
 file_path = os.path.join(script_dir, "destination_file.txt")
 
-newfile = open(file_path, "w")
-
-destinationFile = open(file_path, "w")
-
-print(content)
-destinationFile.write(content)
-
+#If the file dosnt exists make a new. 
+if not os.path.isfile(file_path):
+    newfile = open(file_path, "w")
+    destinationFile = open(file_path, "w")
+    destinationFile.write(content)
+else: 
+    try:
+        file = open(file_path,"w")
+        file.write(content)
+    except (PermissionError) as e:
+        print("Fejl i skrivning")
+        sys.exit("Destinations-filen eksistere allerede og er skrivebeskyttet")
+        
 destinationFile.flush()
 destinationFile.close()
-
-#skrivebestyttelse
-errorname = "IOError"
